Rename selftests package
