+++ src/main/java/org/team1540/robot2019/commands/selftests/SelfTest.java
@@ -1,4 +1,4 @@
-package org.team1540.robot2019.commands.self_tests;
+package org.team1540.robot2019.commands.selftests;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
 import org.apache.log4j.Logger;
